requirements.txt
crewai
langchain
langchain-community
langchain-openai
sentence-transformers
pypdf
fastapi
uvicorn
python-dotenv
pyyaml
pydantic
litellm
psycopg2-binary
pgvector
